refactor(lib): move NrsMap and related functionality to its own nrs_map.rs file
